User Friendly Classroom Planner
A classroom planner that focuses on making the tool more user friendly. This tool changes almost all input to be through a graphical user interface and allows the user to navigate through the different method using a menu. This tool will also aim to add the features from the other classroom planners to combine them into one version.